Small changes on UnfinishedWorkService
